refactor: 股本字段返回原始股数
- 返回 int 类型（原始数据）
- 调用方自行决定如何格式化（亿股/万手等）
- 保持数据精度

+++ stock/models.py
@@ -112,10 +112,10 @@ class BonusHistory:
 class SharesChangeRecord:
     """股本变动记录"""
     chg_date: Optional[str] = None                  # 变动日期
-    total_shares: Optional[float] = None          # 总股本（亿股）
-    float_shares_ashare: Optional[float] = None     # 流通A股（亿股）
-    float_shares_hshare: Optional[float] = None    # 流通H股（亿股）
-    chg_reason: Optional[str] = None               # 变动原因
+    total_shares: Optional[int] = None              # 总股本（股）
+    float_shares_ashare: Optional[int] = None        # 流通A股（股）
+    float_shares_hshare: Optional[int] = None       # 流通H股（股）
+    chg_reason: Optional[str] = None                # 变动原因
 
     def to_dict(self) -> Dict[str, Any]:
         return {k: v for k, v in asdict(self).items() if v is not None}
